requirements.txt
flask
pandas
langchain
Flask-Cors
langchain-openai
python-dotenv
sqlalchemy